replaced test of 25! with test of 18!
rationale: 18! is the largest factorial less than 2^53-1 (the precision of a double). When testing 25!, there could be rounding errors when the implemented solution was different from the solution provided by TOP, which resulted in the test to fail despite having a correct algorithm. Testing with 18! instead of 25! ensures that there will be no such fail.